requirements.txt
setuptools
cython

numpy>=1.12.1
scipy>=0.18.1
six
psutil
pynisher>=0.4.1
ConfigSpace>=0.4.3,<0.5
scikit-learn>=0.18.0
typing
pyrfr>=0.5.0
sphinx
sphinx_rtd_theme
joblib


emcee>=2.1.0
matplotlib>=1.3.1
cma
direct
theano==0.9.0
git+https://github.com/Lasagne/Lasagne.git
git+https://github.com/automl/george.git
git+https://github.com/stokasto/sgmcmc.git
git+https://github.com/automl/HPOlib2.git
nose

keras
tensorflow_gpu==1.7.0
dill